Cleaned up the code

+++ src/compile/mod.rs
@@ -1,5 +1,5 @@
 pub mod linux_x86_64;
 pub mod commands;
 
-pub const MEM_SZ: u32 = 640 * 1000; // 4kb
-pub const STRING_SZ: u32 = 640 * 1000; // 4kb
+pub const MEM_SZ: usize = 640 * 1000; // 4kb
+pub const STRING_SZ: usize = 640 * 1000; // 4kb
